frontend/latex/output: refactor pdf control bar, make controls more compact if frame is narrow
